Touch propagation if a widget can't process a specific event.
Touch acquisition, to avoid re-processing the whole scene tree when a widget already has the event owned.

Touch tolerance of 10 pixels, if someone is trying to tap, he is allowed to move the finger 10 pixels. This should be per-resolution, but it's working right now, so i'll leave it for the moment.

Optimizations, it runs really smooth.